# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IT-Ticket API
  version: 1.1.0
  description: >
    REST-API för IT-Ticket — ett ärendehanteringssystem för intern IT-support på Prefabmästarna.


    **Leverans:** rullande från `main` (en instans per deployment).
    Versionen ovan spårar API-kontraktet, inte appens git-tagg.


    ## API coverage

    This spec documents the full API surface: 160 operations
    across every route file — health/CSRF, auth, tickets, categories,
    contacts, companies, kb, public, users, API keys,
    settings, comments, attachments, shares, links, checklists, checklist
    templates, template fields, tags,
    reports, backup, push notifications, webhooks, and inbound email.


    **Prose reference:** `docs/API.md` — generated from the actual Express
    routers in `server/src/routes/` and kept up to date on route changes. Use
    it as a narrative cross-check alongside this machine-readable spec.


    ## Autentisering

    Två sätt att autentisera, båda via `Authorization`-headern:

    - **Session (JWT):** `Authorization: Bearer <jwt>`. Access-token gäller 15 min;
      en HttpOnly refresh-cookie (path `/api/auth`) förnyar den via
      `POST /api/auth/refresh`. **Muterande anrop (POST/PUT/PATCH/DELETE) kräver
      även CSRF-token** i `x-csrf-token` (hämtas från `GET /api/csrf-token`).
    - **API-nyckel:** `Authorization: Bearer itk_live_...`. Nycklar har scope
      `read`, `read,write` eller — för administrativa endpoints — även `admin`.
...


...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /templates:
    get:
      operationId: templatesList
      tags: [Templates]
      summary: List ticket templates
      description: Any authenticated user. Returns every ticket template with its dynamic fields (empty array for standard templates).
      responses:
        "200":
          description: Ticket templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Template" }
    post:
      operationId: templatesCreate
      tags: [Templates]
      summary: Create a ticket template (admin)
      description: >
        Admin only. `description_template` is required unless `template_type`
        is `dynamic`. When `template_type` is `dynamic`, `fields` (if given)
        is inserted inline as the template's fields in the same request —
        there is no need to call the template-fields endpoints separately at
        creation time.
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateInput" }
      responses:
        "201":
          description: Created template with fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /templates/reorder:
    put:
      operationId: templatesReorder
      tags: [Templates]
      summary: Reorder ticket templates (admin)
      description: >
        Admin only. `ids` is the full list of template ids in the desired
        order; each id's `position` is set to its index in the array. Must be
        registered before `/templates/{id}` in the router so `reorder` is not
        captured as an `:id` path segment.
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: All templates in their new order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Template" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /templates/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: templatesGet
      tags: [Templates]
      summary: Get a single ticket template with its fields
      description: Any authenticated user.
      responses:
        "200":
          description: The template with its fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      operationId: templatesUpdate
      tags: [Templates]
      summary: Update a ticket template (admin)
      description: >
        Admin only. Partial update (each omitted field keeps its current
        value). Does NOT accept `fields` — dynamic-template fields are
        managed exclusively through the `/templates/{templateId}/fields`
        endpoints.
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateUpdateInput" }
      responses:
        "200":
          description: Updated template with its (unchanged) fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: templatesDelete
      tags: [Templates]
      summary: Delete a ticket template (admin)
      description: Admin only. Cascades to the template's fields (`template_fields.template_id` is `ON DELETE CASCADE`).
      security:
        - bearerAuth: []
          csrfToken: []
      responses:
        "200": { $ref: "#/components/responses/Message" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /templates/{templateId}/fields:
    parameters:
      - { name: templateId, in: path, required: true, schema: { type: string } }
    get:
      operationId: templateFieldsList
      tags: [Template fields]
      summary: List a template's dynamic fields
      description: Any authenticated user.
      responses:
        "200":
          description: Fields ordered by position
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TemplateField" }
    post:
      operationId: templateFieldsCreate
      tags: [Template fields]
      summary: Add a dynamic field to a template (admin)
      description: Admin only. Appended at the end (`position` = current max + 1).
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateFieldInput" }
      responses:
        "201":
          description: Created field
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TemplateField" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Session-JWT: `Authorization: Bearer <jwt>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API-nyckel: `Authorization: Bearer itk_live_...` (scope read/write/admin; admin-skyddade endpoints kräver `admin` även när ägaren är administratör). Undantagen från CSRF."
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Krävs på muterande sessions-anrop. Hämtas från `GET /api/csrf-token`.
  responses:
    Error:
      description: Fel
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Message:
      description: OK med meddelande
      content:
        application/json:
          schema:
            type: object
            properties:
              message: { type: string }
...
    Forbidden:
      description: Saknar behörighet (t.ex. admin krävs eller fel scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Hittades inte
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Template:
      type: object
      description: >
        A reusable ticket blueprint used to prefill `POST /tickets`. `standard`
        templates prefill title/description/etc.; `dynamic` templates instead
        carry a set of custom form `fields` used to compose the ticket body.
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        template_type: { type: string, enum: [standard, dynamic] }
        title_template: { type: string }
        description_template: { type: string, description: Empty string for dynamic templates rather than null (column is NOT NULL). }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string, nullable: true }
        notes_template: { type: string, nullable: true }
        solution_template: { type: string, nullable: true }
        position: { type: integer }
        created_by: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        fields:
          type: array
          items: { $ref: "#/components/schemas/TemplateField" }
    TemplateInput:
      type: object
      required: [name, title_template]
      properties:
        name: { type: string }
        description: { type: string }
        template_type: { type: string, enum: [standard, dynamic], default: standard }
        title_template: { type: string }
        description_template: { type: string, description: Required unless template_type is dynamic. }
        priority: { type: string, enum: [low, medium, high, critical], default: medium }
        category_id: { type: string, nullable: true }
        notes_template: { type: string, nullable: true }
        solution_template: { type: string, nullable: true }
        fields:
          type: array
          description: Only used when template_type is dynamic; inserted inline for this template.
          items: { $ref: "#/components/schemas/TemplateFieldInput" }
    TemplateUpdateInput:
      type: object
      description: All fields optional; each omitted field keeps its current value. Fields (custom form fields) cannot be set here.
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        template_type: { type: string, enum: [standard, dynamic] }
        title_template: { type: string }
        description_template: { type: string }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string, nullable: true }
        notes_template: { type: string, nullable: true }
        solution_template: { type: string, nullable: true }
    TemplateField:
      type: object
      description: A single custom form field belonging to a dynamic Template.
      properties:
        id: { type: string }
        template_id: { type: string }
        field_name: { type: string }
        field_label: { type: string }
        field_type: { type: string, enum: [text, textarea, number, select, date, checkbox] }
        placeholder: { type: string, nullable: true }
        default_value: { type: string, nullable: true }
        required:
          type: integer
          enum: [0, 1]
          description: >
            SQLite boolean stored/returned as 0 or 1. Unlike ticket-checklist
            items, this route does not map it to a JSON boolean.
        options:
          type: string
          nullable: true
          description: >
            JSON-encoded array of option strings, used when field_type is
            select. Stored/returned as the raw serialized string, not parsed.
        position: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TemplateFieldInput:
      type: object
      required: [field_name, field_label, field_type]
      description: >
        Used for both create (all required fields must be present) and
        update (all fields optional — omitted ones keep their current value).
      properties:
        field_name: { type: string }
        field_label: { type: string }
        field_type: { type: string, enum: [text, textarea, number, select, date, checkbox] }
        placeholder: { type: string, nullable: true }
        default_value: { type: string, nullable: true }
        required:
          type: boolean
          description: Coerced to 0/1 on write.
        options:
          description: >
            Options for a select field. The route re-serializes whatever is
            sent with JSON.stringify before storing it — send a plain array
            of strings here, not a pre-encoded JSON string.
          type: array
          items: { type: string }
        position: { type: integer }
...